Move connectivity-related classes to a dedicated `connectivity` package for better organization and clarity.
