+++ app/src/main/kotlin/com/addhen/livefront/connectivity/NetworkConnectivityRepository.kt
@@ -1,4 +1,4 @@
-package com.addhen.livefront
+package com.addhen.livefront.connectivity
 
 import android.content.Context
 import dagger.hilt.android.qualifiers.ApplicationContext
