+++ app/src/main/kotlin/com/addhen/livefront/screen/githubrepolist/GithubRepoListScreen.kt
@@ -46,7 +46,7 @@ import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.collectAsLazyPagingItems
 import coil.compose.AsyncImage
 import coil.request.ImageRequest
-import com.addhen.livefront.ConnectionState
+import com.addhen.livefront.connectivity.ConnectionState
 import com.addhen.livefront.R
 import com.addhen.livefront.data.model.GithubRepo
 import com.addhen.livefront.formatStars
